Now, everything is casted to float if it's possible. Comparisons not working, probably because I assumed that it would be as simple as comparing integers

+++ examples/fixed/fixed.ino
@@ -12,34 +12,35 @@
 using namespace espmath;
   
 typedef struct Parameters{
-    fixed target = 8.0f;
-    fixed observed = 18.0f;
-    fixed ambient = 20.0f;
-    fixed heatsink = 26.0f;
-    fixed cooling = 0.0f;
+    fixed target = 8;
+    fixed observed = 18;
+    fixed ambient = 20;
+    fixed heatsink = 26;
+    fixed cooling = 0;
     char energy = 60;
     const uint16_t maxmV = 8000;
     const uint32_t expectTime = 1800;
     const char energyTh = 70;
-    const fixed heatsinkTh = 32.0f;
+    const fixed heatsinkTh = 32;
 }param;
 
 void regulator(Parameters* p)
 {
-  fixed X0 = (p->observed - p->target) / (abs((float)p->observed) + abs((float)p->target));
-  fixed X1 = 1.0f - ((float)((p->expectTime - 360) / p->expectTime));
-  fixed X2 = (p->ambient - p->target) / (abs((float)p->ambient) + abs((float)p->target));
+  fixed X0 = (p->observed - p->target) / (abs(p->observed) + abs(p->target));
+  fixed X1 = 1 - ((float)((p->expectTime - 360) / p->expectTime));
+  fixed X2 = (p->ambient - p->target) / (abs(p->ambient) + abs(p->target));
   fixed X3 = (p->heatsink - p->ambient) / (p->heatsinkTh - p->ambient);
-  fixed X4 = p->energy > p->energyTh ? 1.0f : p->energy / 100.0f;
+  fixed X4 = p->energy > p->energyTh ? 1 : p->energy / 100;
   X4 *= X4;
   fixed X5 = p->cooling - p->target;
   
-  debug.print("X0*C0: " + String(float(X0*C0)));
-  debug.print("X1*C1: " + String(float(X1*C1)));
-  debug.print("X2*C2: " + String(float(X2*C2)));
-  debug.print("X3*C3: " + String(float(X3*C3)));
-  debug.print("X4*C4: " + String(float(X4*C4)));
+  debug.print("X0*C0: " + String(X0*C0));
+  debug.print("X1*C1: " + String(X1*C1));
+  debug.print("X2*C2: " + String(X2*C2));
+  debug.print("X3*C3: " + String(X3*C3));
+  debug.print("X4*C4: " + String(X4*C4));
   
+  // Not working. Need to be reviewed
   // if ((!X0 && !X2) || X0 < 0.f || X3 >= 1.f || X5 > 0.f)
   // {
   //   debug.print("Shutdown!");
@@ -50,7 +51,7 @@ void regulator(Parameters* p)
   fixed activF = tanh(x);
   
   debug.print(activF);
-  debug.print(((activF + 1.f) / 2.f) * (p->maxmV / 1000.f));
+  debug.print(((activF + 1.) / 2.) * (p->maxmV / 1000.0f));
 }
 
 void setup()
